refactor: extract statistics cards list into separate component with horizontal scroll
